+++ src/artist/artist.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import dbManager from 'src/_db/dbManager';
 import { IArtistData } from 'src/_db/Artists';
+import { isUUID } from '../utils/isUUID';
 
 @Injectable()
 export class ArtistService {
@@ -19,10 +20,22 @@ export class ArtistService {
   }
 
   findOne(id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.artistExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     return dbManager.getArtistById(id);
   }
 
   update(id: string, updateArtistDto: UpdateArtistDto) {
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.artistExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
     let artistData: IArtistData = {
       name: updateArtistDto.name ?? null,
       grammy: updateArtistDto.grammy ?? null,
@@ -31,6 +44,13 @@ export class ArtistService {
   }
 
   remove(id: string) {
-    return dbManager.deleteArtist(id);
+    if (!isUUID(id)) {
+      throw new HttpException('Wrong id', HttpStatus.BAD_REQUEST)
+    }
+    if (!dbManager.artistExists(id)) {
+      throw new HttpException('Unknown record', HttpStatus.NOT_FOUND)
+    }
+    dbManager.deleteArtist(id);
+    dbManager.deleteArtistRelation(id);
   }
 }
